removed stored procedure inception
